Added dns record various improvements with types, working with parsing the response

+++ dns-resolver/src/DNSQuestion.cpp
@@ -1,6 +1,6 @@
 #include "DNSQuestion.hpp"
 
-DNSQuestion::DNSQuestion(int type, int klass, std::vector<uint8_t> name) {
+DNSQuestion::DNSQuestion(uint16_t type, uint16_t klass, std::vector<uint8_t> name) {
     this->type = type;
     this->klass = klass;
     this->name = name;
